🐞 Fix: DesignSystem 영역에서 Domain에 접근하려다 생긴 오류 수정
Domain에 Common 과 DesignSystem 의존성 추가
