+++ Projects/Data/Sources/DTO/School/SchoolDataDTO.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2024 ATeen. All rights reserved.
 //
 
+import Common
 import Domain
 import Foundation
