+++ Projects/Data/Sources/Service/SearchSchoolServiceImp.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2024 ATeen. All rights reserved.
 //
 
+import Common
 import Domain
 import Foundation
