+++ Projects/DesignSystem/Derived/Sources/TuistBundle+DesignSystem.swift
@@ -4,7 +4,7 @@
 import Foundation// MARK: - Swift Bundle Accessor for Frameworks
 private class BundleFinder {}
 extension Foundation.Bundle {
-/// Since DesignSystem is a dynamic framework, the bundle for classes within this module can be used directly.
+/// Since DesignSystem is a framework, the bundle for classes within this module can be used directly.
 static let module = Bundle(for: BundleFinder.self)
 }// MARK: - Objective-C Bundle Accessor
 @objc
