+++ Projects/DesignSystem/Sources/Components/Cell/SearchSchoolResultTableViewCell.swift
@@ -6,7 +6,6 @@
 //
 
 import Common
-import Domain
 import UIKit
 
 public final class SearchSchoolResultTableViewCell: UITableViewCell {
