+++ Projects/DesignSystem/Sources/Components/Cell/TeenTableViewCell.swift
@@ -9,7 +9,6 @@
 import SnapKit
 
 import Common
-import Domain
 import UIKit
 
 public final class TeenTableViewCell: UITableViewCell {
