+++ Projects/DesignSystem/Sources/Components/View/CustomQuestionView.swift
@@ -9,7 +9,6 @@
 import SnapKit
 
 import Common
-import Domain
 import UIKit
 
 public class CustomQuestionView: UIView {
