+++ Projects/Domain/Project.swift
@@ -11,5 +11,7 @@ import ProjectDescriptionHelpers
 let project = Project.makeProject(
     name: "Domain",
     moduleType: .dynamicFramework,
-    dependencies: []
+    dependencies: [
+        .common
+    ]
 )
