+++ Projects/Domain/Sources/Entity/Request/MyPage/MyPageEditRequest.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2024 ATeen. All rights reserved.
 //
 
+import Common
 import Foundation
 
 public struct MyPageEditRequest {
