+++ Projects/Domain/Sources/Entity/Response/User/UserDetailResponse.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2024 ATeen. All rights reserved.
 //
 
+import Common
 import Foundation
 
 public struct UserDetailResponse: Decodable {
