+++ Projects/Domain/Sources/Entity/Response/User/UserFindResponse.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2024 ATeen. All rights reserved.
 //
 
+import Common
 import Foundation
 
 public struct UserFindResponse: Decodable {
@@ -29,30 +30,3 @@ public struct UserData: Decodable {
     }
 }
 
-public struct User: Decodable {
-    public let id: Int
-    public let uniqueId: String
-    public let profileImage: String?
-    public let nickName: String
-    public let location: String
-    public let schoolName: String
-    public var likeStatus: Bool
-    
-    public init(
-        id: Int,
-        uniqueId: String,
-        profileImage: String?,
-        nickName: String,
-        location: String,
-        schoolName: String,
-        likeStatus: Bool
-    ) {
-        self.id = id
-        self.uniqueId = uniqueId
-        self.profileImage = profileImage
-        self.nickName = nickName
-        self.location = location
-        self.schoolName = schoolName
-        self.likeStatus = likeStatus
-    }
-}
