+++ Projects/Domain/Sources/UseCase/SignUseCase.swift
@@ -7,6 +7,7 @@
 //
 
 import Combine
+import Common
 import Foundation
 
 public protocol SignUseCase {
