+++ Projects/Domain/Sources/UseCaseImplementations/SignUseCaseImp.swift
@@ -7,6 +7,7 @@
 //
 
 import Combine
+import Common
 import Foundation
 
 public struct SignUseCaseImp: SignUseCase {
