+++ Projects/Feature/ReportFeature/Derived/InfoPlists/ReportFeature-Info.plist
@@ -17,7 +17,7 @@
 	<key>CFBundleShortVersionString</key>
 	<string>1.0.0</string>
 	<key>CFBundleVersion</key>
-	<string>2024.09.25.02.30</string>
+	<string>2024.09.12.15.01</string>
 	<key>HOST_VALUE</key>
 	<string>$(HOST_VALUE)</string>
 </dict>
